requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
Werkzeug
datetime
pymongo
python-dotenv
pandas
numpy
requests
email_validator
os